+++ CHANGELOG.md
@@ -1,6 +1,43 @@
 # Changelog 
 
-[Unreleased changes](https://github.com/rapidez/blade-directives/compare/0.1.0...master)
+[Unreleased changes](https://github.com/rapidez/blade-directives/compare/0.4.2...master)
+## [0.4.2](https://github.com/rapidez/blade-directives/releases/tag/0.4.2) - 2024-02-29
+
+### Fixed
+
+- Fix method_exists check (#9)
+- Use Arrayable on OptionalDeep (#10)
+
+## [0.4.1](https://github.com/rapidez/blade-directives/releases/tag/0.4.1) - 2024-02-27
+
+### Fixed
+
+- Use CanBeEscapedWhenCastToString (#8)
+
+## [0.4.0](https://github.com/rapidez/blade-directives/releases/tag/0.4.0) - 2024-02-22
+
+### Added
+
+- Allow attempted casting to boolean and array (#7)
+
+## [0.3.1](https://github.com/rapidez/blade-directives/releases/tag/0.3.1) - 2024-02-21
+
+### Fixed
+
+- Handle passing optionalDeep into optionalDeep (#6)
+
+## [0.3.0](https://github.com/rapidez/blade-directives/releases/tag/0.3.0) - 2024-01-24
+
+### Added
+
+- optionalDeep helper (#4)
+
+## [0.2.0](https://github.com/rapidez/blade-directives/releases/tag/0.2.0) - 2023-12-12
+
+### Added
+
+- Markdown directive (#1)
+
 ## [0.1.0](https://github.com/rapidez/blade-directives/releases/tag/0.1.0) - 2023-07-21
 
 ### Added